+++ device/api/umd/device/blackhole_implementation.h
@@ -10,7 +10,7 @@
 #include <stdexcept>
 
 #include "umd/device/architecture_implementation.h"
-#include "umd/device/tlb.h"
+#include "umd/device/types/tlb.h"
 
 namespace tt::umd {
