+++ device/api/umd/device/cluster.h
@@ -18,12 +18,12 @@
 #include "tt_soc_descriptor.h"
 #include "tt_xy_pair.h"
 #include "umd/device/chip/chip.h"
-#include "umd/device/tlb.h"
 #include "umd/device/tt_device/tt_device.h"
 #include "umd/device/tt_io.hpp"
 #include "umd/device/types/arch.h"
 #include "umd/device/types/cluster_descriptor_types.h"
 #include "umd/device/types/cluster_types.h"
+#include "umd/device/types/tlb.h"
 
 using TLB_DATA = tt::umd::tlb_data;
